feat: Add optional path curve sampling for planned pedestrian trajectories
- Introduced `curve_offset_m` and `curve_sample_spacing_m` parameters to adjust baseline path geometry with deterministic curves.
- Implemented dynamic path smoothing with validation for obstacle clearance, falling back to raw polyline if violated.
- Enhanced `FEpisodePedestrianPlanBuilder` to handle path curve parameters and generate adjusted plans.
- Added tests for curve sampling, hash changes, and runtime validation.
- Updated JSON guide and documentation with new movement fields for planned pedestrians.

+++ Source/ProtoRobotSim/Private/Episode/Components/EpisodePedestrianRuntimeComponent.cpp
@@ -113,7 +113,7 @@ void UEpisodePedestrianRuntimeComponent::TickComponent(float deltaTime, ELevelTi
 		return;
 	}
 
-	const double deltaSeconds = static_cast<double>(deltaTime);
+	const double deltaSeconds = deltaTime;
 	ActiveTimeSeconds += FMath::Max(deltaSeconds, 0.0);
 	UpdateRobotDistanceMetrics();
